+++ src/main/java/cubicchunks/asm/mixin/core/client/MixinWorldProvider.java
@@ -21,7 +21,7 @@
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *  THE SOFTWARE.
  */
-package cubicchunks.asm.mixin.core.common;
+package cubicchunks.asm.mixin.core.client;
 
 import cubicchunks.world.ICubicWorld;
 import cubicchunks.world.NotCubicChunksWorldException;
@@ -36,6 +36,9 @@ import net.minecraft.world.DimensionType;
 import net.minecraft.world.World;
 import net.minecraft.world.WorldProvider;
 import net.minecraft.world.chunk.IChunkGenerator;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
+
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
@@ -53,66 +56,23 @@ public abstract class MixinWorldProvider implements ICubicWorldProvider {
 
     @Shadow protected World world;
 
-    @Shadow protected boolean hasNoSky;
-
-    @Shadow public abstract int getDimension();
-
-    @Shadow public abstract DimensionType getDimensionType();
-
-    @Shadow public abstract IChunkGenerator createChunkGenerator();
-
     /**
-     * @reason return the real world height instead of hardcoded 256
-     * @author Barteks2x
+     * Return Double.NaN to remove void fog and fix night vision potion below Y=0.
+     * <p>
+     * In EntityRenderer.updateFogColor entity Y position is multiplied by value returned by this method.
+     * <p>
+     * If this method returns any real number - then the void fog factor can be <= 0. But if this method returns NaN -
+     * the result is always NaN. And Minecraft enables void fog only of the value is < 1. And since any comparison with
+     * NaN returns false - void fog is effectively disabled.
      */
-    // @Overwrite() - overwrite doesn't support unobfuscated methods
-    public int getHeight() {
-        return cubicWorld().getMaxHeight();
-    }
-
-    /**
-     * @reason return the real world height instead of hardcoded 256
-     * @author Barteks2x
-     */
-    // @Overwrite() - overwrite doesn't support unobfuscated methods
-    public int getActualHeight() {
-        // only give the real value for overworld, mods may use it scan height start in their teleporter code
-        return hasNoSky ? 128 : getDimension() == 0 ? getHeight() : 256;
-    }
-
-    @Nullable @Override public ICubeGenerator createCubeGenerator() {
-        if (!cubicWorld().isCubicWorld()) {
-            throw new NotCubicChunksWorldException();
-        }
-        if (this.getDimensionType() == DimensionType.OVERWORLD) {
-            return ((ICubicWorldType) cubicWorld().getWorldType()).createCubeGenerator(cubicWorld());
-        }
-        return new VanillaCompatibilityGenerator(this.createChunkGenerator(), cubicWorld());
-    }
-
-    @Inject(method = "getRandomizedSpawnPoint", at = @At(value = "HEAD"), cancellable = true, remap = false)
-    private void findRandomizedSpawnPoint(CallbackInfoReturnable<BlockPos> cir) {
+    @Inject(method = "getVoidFogYFactor", at = @At(value = "HEAD"), cancellable = true)
+    private void getVoidFogYFactor_injectReplace(CallbackInfoReturnable<Double> cir) {
         if (cubicWorld().isCubicWorld()) {
-            cir.setReturnValue(new SpawnPlaceFinder().getRandomizedSpawnPoint(cubicWorld()));
+            cir.setReturnValue(Double.NaN);
             cir.cancel();
         }
     }
-
-/*
-    @Override
-    @Deprecated
-    public boolean canDropChunk(int x, int z) {
-        return true;
-    }
-
-    @Override
-    public boolean canCoordinateBeSpawn(int x, int z) {
-        //TODO: DONT USE WORLD.getGroundAboveSeaLevel()
-        BlockPos blockpos = new BlockPos(x, 0, z);
-        return this.world.getBiome(blockpos).ignorePlayerSpawnSuitability() ||
-                this.world.getGroundAboveSeaLevel(blockpos).getBlock() == Blocks.GRASS;
-    }
-*/
+    
     private ICubicWorld cubicWorld() {
         return (ICubicWorld) world;
     }
